+++ packages/integrations/markdoc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @astrojs/markdoc
 
+## 0.15.5
+
+### Patch Changes
+
+- Updated dependencies [[`4d16de7`](https://github.com/withastro/astro/commit/4d16de7f95db5d1ec1ce88610d2a95e606e83820)]:
+  - @astrojs/internal-helpers@0.7.2
+  - @astrojs/markdown-remark@6.3.6
+
 ## 0.15.4
 
 ### Patch Changes